Move snapshot helpers into its own file
